backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
python-dotenv==1.0.1
psycopg[binary]==3.2.3
qdrant-client==1.12.1
neo4j==5.27.0
wntr==1.4.0
scikit-learn==1.7.2
